Wheat-Counter-Docker-Flask-App /Wheat-Backend/flask/requirements.txt
Click==7.0
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
uWSGI==2.0.18
Werkzeug==0.16.0
numpy==1.15.1 
segmentation-models-pytorch==0.1.0
albumentations==0.4.5
opencv-python==4.1.2.30
Pillow==5.4.1